🔧 chore(node.js.yml): add prettier:check script to the workflow 📝 docs(CHANGELOG.md): update changelog for version 1.0.12 📝 docs(CONTRIBUTING.md): update coding guidelines, pull request policies, and commitlint configuration 🔧 chore(commitlint.config.js): update commitlint configuration to extend commitlint-config-gitmoji and disable some rules 🖼️ feat(images): add conventional-commits.png 📦 chore(package.json): update version to 1.0.12 and update commit script to use oco instead of manually formatting the commit message
The prettier:check script is added to the workflow to ensure code formatting consistency. The CHANGELOG.md is updated to include the changes made in version 1.0.12. The CONTRIBUTING.md file is updated to provide guidelines for naming conventions, development commands, changesets, conventional commits, Husky, Commitlint, and OpenCommit. The commitlint.config.js file is updated to extend commitlint-config-gitmoji and disable some rules. The conventional-commits.png image is added. The package.json file is updated to use oco for formatting commit messages and to update the version to 1.0.12.