🔧 fix(node.js.yml): rename workflow from "Node.js CI" to "Validation" for better clarity ✨ feat(node.js.yml): update node versions in the matrix to [18.x, 20.x] to align with supported versions The workflow name has been changed from "Node.js CI" to "Validation" to better reflect its purpose. The node version matrix has been updated to include versions 18.x and 20.x, which are the supported versions. This ensures that the workflow runs on the specified node versions for testing and building the source code.
🔧 fix(package.json): update version from 1.0.18 to 1.3.0 to reflect the latest changes
The package version has been updated from 1.0.18 to 1.3.0 to reflect the latest changes and improvements made to the codebase. This allows for better tracking and identification of the changes made in the project.

🔧 fix(index.ts): refactor code to use OpenAI API for log analysis and error resolution
✨ feat(index.ts): add functionality to analyze build logs and generate hints for issue resolution using OpenAI API
The code in index.ts has been refactored to utilize the OpenAI API for log analysis and error resolution. The Azure DevOps task now analyzes the provided build logs, identifies the root cause of the issues, and generates clear and structured hints to resolve the problems. This integration with OpenAI enhances the pipeline debugging process and empowers developers to optimize their pipeline workflows.

🔧 fix(tslint.json): disable specific tslint rules to allow console, var requires, and shadowed variables
The tslint.json file has been updated to disable specific tslint rules. The "no-console", "no-var-requires", and "no-shadowed-variable" rules have been disabled to allow the usage of console statements, var requires, and shadowed variables in the codebase. This change improves the flexibility and readability of the code.